backend/requirements.txt
pikepdf>=8.0.0
werkzeug>=2.0.0
Flask
gunicorn
Flask-CORS
Pillow
scikit-learn
numpy
